requirements.txt
fastapi
uvicorn
anthropic
pdfplumber
python-docx
httpx
pydantic
python-multipart
python-dotenv